requirements.txt
flask==3.0.2
sympy==1.12
python-dotenv==1.0.1
google-generativeai==0.2.2
asgiref==3.7.2